Fine scale

YiyuanStudentProject.generate_rve_grid — Function

generate_rve_grid(; ϕ::Real, d::Real, meshsize::Real, dx::NTuple{3,Real}=(1.0,1.0,1.0))
generate_rve_grid(spheres::Vector{BubbleBath.Sphere{3}}; meshsize::Real, dx::NTuple{3,Real}=(1.0,1.0,1.0))

Return a Grid with spherical inclusions for a 3D RVE in desired meshsize. Two cell sets are included in the generated grid:

"particles": Cells within the spherical inclusions.

"matrix": Cells outside the spheres (background material).

Variation 1

The grid will be generated by giving the volume fraction of the spherical inclusions ϕ (0 ≤ ϕ ≤ 1), the average diameter of the spheres d, and the meshsize as well as the RVE size dx in 3D, which by default is (1.0, 1.0, 1.0).

Variation 2

In this variation a vector of pre-defined spheres to be embedded in the grid sphere is given instead of the volume fraction and the average diameter of the sphere. The meshsize and the RVE size dx are still needed arguments in this case.

YiyuanStudentProject.prepare_setup — Function

prepare_setup(rve::RVE{dim}) where {dim}

Return a RVESetup struct for the element-wise assembly and time stepping.

Arguments:

-rve: RVE

Implementation Details:

The interpolation ip is defined by passing the corresponding refshape using the function _get_ref_shape(Val(dim)).

After the 'DofHandler' is created based on the grid from the argument rve, unknown fields can be added to it using add! By calling close! to finalize the construction.

Cellvallues are created for each discrete fields respectively passing quadrature rule with 3 integration points.

Number of base function for each unknown fields is defined using getnbasefunctions.

Calling function add_bc! to initialize the boundary conditions and associate this to the $Dofhandler$.

Local stiffness and mass matrices are initialized. subarrays for locating the corresponding field interaction in element-wise by calling the macro @view.

A $NamedTuple$ with the PhaseSetups is then prepared for particles P and matrix M.

Global stiffness, mass, and system matrices is initialized using allocate_matrix for a sparse matrix pattern.

Initialize the gloable system vector and the solution vectors for both current and next time step.

Get the total volume of the RVE.

YiyuanStudentProject.add_bc! — Function

add_bc!(ch::ConstraintHandler, grid::Grid{3}, load::LoadCase{3})

Add a dirichlet boundary condition for the unknown fields respectively on the ∂Ω part of the boundary. ∂Ω is defined by the collection of periodic facets from the grid.

YiyuanStudentProject.assemble_K_M_f! — Function

    assemble_K_M_f!(setup::RVESetup)

Return the assembled mass matrix M and stiffness matrix K.

Arguments:

setup: A preconfigured object that defined for RVE assembly. Fields from setup are used in this function:

phasesetups: A preconfigured object type that defined for elementweise assembly,

M: Initialized global mass matrix to be assembled,

K: Initialized global stiffness matrix to be assembled,

f: Initialized global right hand side vector to be assembled.

Implementation Details:

Create a CSCAssembler for both K, f, and M.

Do the whole assembly for all the phases: for all cells update the cell values for all the unknown fields and fill zeros to the do element assembly, return assembled Kₑ, Mₑ and fₑ.

YiyuanStudentProject.assemble_element! — Function

    assemble_element!(setup::PhaseSetup)

Return the assembled mass matrix for coupling between the chemical potantial and the concentration fields as well as the stiffness matrix.

Arguments:

setup: a preconfigured object type that defined for elementweise assembly.

Implementation Details:

For each quadrature point the element volume is computed. Furthermore for each base function in corresponding field, evaluate the shape function for the test function. The nodal value of a certain unknown field is then computed for each base funtion in that field.

Then the coupling subarrays are computed as:

Kₑuu = ∫(δNϵi ⊡ E ⊡ Nϵj) * dΩ

Kₑuc = ∫-(δNϵi ⊡ E ⊡ (αᶜʰ*(Ncj - cʳᵉᶠ))) * dΩ

Kₑcu = ∫(δNci * (αᶜʰ ⊡ E ⊡ Nϵj))* dΩ

Kₑcc = ∫-(δNci * (k(Ncj-cʳᵉᶠ) + αᶜʰ ⊡ E ⊡(αᶜʰ(Ncj - cʳᵉᶠ))))* dΩ

Kₑcμ = ∫(δNci * (Nμj - μʳᵉᶠ)) * dΩ

Kₑμμ = ∫(δN∇μi ⋅ M ⋅ N∇μj) * dΩ

Mₑμc = ∫(Ncj * δNμi) * dΩ

where:

Nxj: Shape function of field x

δNxi: Shape function for x field test function

N∇xj: Gradient of shape function of field x

δN∇xi: Gradient of shape function for test function of field x

dΩ: Determinant of the Jacobian times the quadrature weight

YiyuanStudentProject.compute_time_step! — Function

compute_time_step!(setup::RVESetup{dim}, load::LoadCase{dim}, Δt)

Compute the solution for the next time step in a Representative Volume Element (RVE) simulation using an implicit time integration scheme.

Arguments

setup: RVESetup{dim}: The setup object containing parameters for the RVE simulation

load: LoadCase{dim}: The load case specifying boundary conditions and external loads for the current time step.

Δt: The time step size.

Implementation Details

This function uses Crank-Nicolson Method to compute the solution in an implicit time stepping scheme for solving the linear system in the next time step:

The function applies boundary conditions to the system matrix and force vector using a constraint handler before solving for the next time step solutions.

YiyuanStudentProject.solve_time_series — Function

solve_time_series(rve::RVE{dim}, load::LoadCase{dim};  Δt=0.25, t_total=1) where {dim}

For visualizing the fine scale problem results. Compute the results in a NamedTuple with fields t total time and a solution vector series contain u displacement, μ chemical potantial, and c concentration for the whole time series with a certain time step width.

Arguments

rve: Object for solving function prepare_setup,

load: Object LoadCase.

YiyuanStudentProject.plot_grid — Function

plot_grid(grid::Grid{3})

Ruturn a Makie.Figure to visualize a 3D finite element grid using the Makie plotting library.

Arguments:

grid::Grid{3}: A 3D grid object representing the finite element mesh to be visualized.

Implementation Details:

A plotable mesh is generated using _prepare_plotable_mesh and input grid

Sets up a 3D axis Axis3 with an equal aspect ratio and a title undeformed grid.

Renders the grid as a solid mesh Makie.mesh! with a light blue color.

Overlays the grid's edges as a wireframe Makie.wireframe! with black edges.

YiyuanStudentProject.animate_result — Function

animate_result(res::NamedTuple, setup::RVESetup{dim}, file_name::String="Myresult.mp4", n::Number)

Return an animation showing the evolution of the solution for a 3D RVE simulation.

Arguments:

res: A NamedTuple containing the simulation results

setup: The setup object for the RVE simulation, containing: grid and dh fields

file_name: The path and name of the output animation file (default: "Myresult.mp4")

Implementation Details:

Initialize a Makie figure with a size of (1200, 800).

Generate observables for t and a using _prepare_plots!.

Record frames by iterating over the indices of t and updates the observables accordingly.

Save the animation as an MP4 file.