Macro Scale

YiyuanStudentProject.prepare_macro_setup — Function

prepare_macro_setup(grid::Grid{dim,C}, rvesetup::RVESetup{dim}, rve::RVE{dim}, Δt, bc::MacroBCParams) where {dim,C}

Return a SolveSetup object for the element-wise assembly and solving the time dependent macro scale problem.

A initialized buffer data is created with the object GaussPointData at each integration point.

The object PhaseSetup is created based on the given macro scale grid grid, the boundary condition MacroBCParams with the buffer data.

With necessary parameters, initialized global matrix and vectors, and PhaseSetup the SolveSetup is computed.

Two unknown fields deformation u, chemical potential μ are prescribed for all cells.

YiyuanStudentProject.add_macro_bc! — Function

add_macro_bc!(ch::ConstraintHandler, grid::Grid{3}, Δt, bc::MacroBCParams)

Add a time dependent dirichlet boundary condition for the macro scale unknown fields respectively on the possible facet ∂Ω1 and ∂Ω2.

Apply time dependent boundary condition on both unknown fields u and μ.

YiyuanStudentProject.assemble_macro_K! — Function

    assemble_macro_K!(setup::SolveSetup{dim}, Δt) where {dim}

Return the whole $SolveSetup$ with assembled stiffness matrix K.

The assembly of all cells is performed using the given input RVESetup.

YiyuanStudentProject.assemble_macro_element! — Function

    assemble_macro_element!(setup::AssemblySetup{dim}, Δt, gpdata::Vector{GaussPointData{dim}}) where {dim}

Return the assembled stiffness matrix Kₑ.

New variationally consistent macro scale fileds σ̄, ċ, ċ₂, and j̄ are computed using the corresponding GaussPointData{dim} and LoadCase, which consists of the quadrature point values ε̄, μ̄, ζ̄.

Different contributions of Kₑ are computed with these updated macro scale fileds and necessary assembly setups provided by the input AssemblySetup.

YiyuanStudentProject.solve_macro_problem — Function

solve_macro_problem(grid::Grid{dim}, rvesetup::RVESetup{dim}, rve::RVE{dim}, bc::MacroBCParams; Δt=0.1, t_total=1) where {dim}

Return results throughout the time stepping process for visualizing the fine scale problem results.

The macro scale problem is prepared, assembled and solved over a time series using the given grid, RVESetup, RVE, MacroBCParams, t_total, Δt . By default t_total is set to 1 and Δt to 0.1.

Results on both macro scale and the reference RVE are stored in a NamedTuple with fields total time t and solution vector a respectively.

The time dependent boundary condition is updated to the global stiffness matrix and force vector as well as the solution vector.

Using solver from $Pardiso.jl$ for eventually ill conditioned system matrices.

YiyuanStudentProject.animate_macro_result — Function

animate_macro_result(res::NamedTuple, setup::SolveSetup{dim}; file_name ="Myresult.mp4", kwargs...) where {dim}

Return an mp4 file with name file_name, which by default is "Myresult.mp4".

This shows the evolution of the solution res for a macro scale simulation of a multi scale problem based on input SolveSetup.

Further arguments like scale for visibel deformation can be used.

YiyuanStudentProject._prepare_plots! — Function

_prepare_plots!(pos, res::NamedTuple, setup::RVESetup{dim}; scale::Real=1.0, titlestart::String="RVE response") where {dim}

Sets up the necessary plots with a specified layout position pos and observables tᵒᵇˢ and aᵒᵇˢ to visualize the simulation results res of a RVE problem RVESetup: deformation u, chemical potential μ, concentration c.

Optional arguments include a scaling factor scale for visualizing the deformation and a title titlestart. By default the scale is set to 1.0 and titlestart to RVE response.

_prepare_plots!(pos, res::NamedTuple, setup::SolveSetup{dim}; scale::Real=1.0, titlestart::String="macroscale solution") where {dim}

Sets up the necessary plots with a specified layout position pos and observables tᵒᵇˢ and aᵒᵇˢ to visualize the simulation results res of a multi scale problem SolveSetup: deformation u, chemical potential μ.

Optional arguments include a scaling factor scale for visualizing the deformation and a title titlestart. By default the scale is set to 1.0 and titlestart to macroscale solution.

YiyuanStudentProject.animate_combined_result — Function

animate_combined_result(res::NamedTuple, resᵣᵥₑ::NamedTuple, setup::SolveSetup{dim}; file_name ="Myresult.mp4", scale=1.0) where {dim}

Return an mp4 file with name file_name, which by default is "Myresult.mp4".

This shows the evolution of the macro scale solution res and the RVE solution at a corresponding quadrature point cell[1]qp[1] for a multi scale simulation based on input SolveSetup.

Further arguments like scale for visibel deformation can be used, which by default is set to 1.0.