Sub Scale

YiyuanStudentProject.generate_rve_grid — Function

generate_rve_grid(; ϕ::Real, d::Real, meshsize::Real, dx::NTuple{3,Real}=(1.0,1.0,1.0))
generate_rve_grid(spheres::Vector{BubbleBath.Sphere{3}}; meshsize::Real, dx::NTuple{3,Real}=(1.0,1.0,1.0))

Return a Grid with spherical inclusions for a 3D RVE with defined meshsize. Two cell sets are included in the generated grid:

"particles": Cells within the spherical inclusions.

"matrix": Cells outside the spheres (background material).

Variation 1

The grid will be generated by giving the volume fraction of the spherical inclusions ϕ (0 ≤ ϕ ≤ 1), the average diameter of the spheres d, and the meshsize as well as the RVE size dx in 3D, which by default is (1.0, 1.0, 1.0). The RVE center will be at the coordinate system origin 0.0, 0.0, 0.0.

Variation 2

In this variation a vector of pre-defined spheres to be embedded in the grid sphere is given instead of the volume fraction and the average diameter of the sphere. The meshsize and the RVE size dx are still needed arguments in this case.

YiyuanStudentProject.prepare_setup — Function

prepare_setup(rve::RVE{dim}) where {dim}

Return a RVESetup object for the element-wise assembly and time stepping.

With the object RVE as argument, the RVESetup is created by constructing PhaseSetup objects for material phase particles P and matrix M, initializing necessary global matrices and vectors, and the total volume of RVE.

Three unknown fields deformation u, chemical potential μ and ion concentration c are prescribed for all cells over both material phases.

YiyuanStudentProject.add_bc! — Function

add_bc!(ch::ConstraintHandler, grid::Grid{3}, load::LoadCase{3})

Add a Dirichlet boundary condition for the unknown fields respectively on the ∂Ω part of the boundary. ∂Ω is defined by the collection of facets from the grid. Boundary condition values are given by objectLoadCase.

YiyuanStudentProject.assemble_K_M_f! — Function

assemble_K_M_f!(setup::RVESetup)

Return the assembled stiffness matrix K, mass matrix M and right hand side vector f.

The assembly of all cells across all material phases is performed using the given input RVESetup.

YiyuanStudentProject.assemble_element! — Function

assemble_element!(setup::PhaseSetup)

Return the assembled local stiffness and mass matrices (Kₑ, Mₑ), the local right hand side vector (fₑ).

Different contributions of Kₑ, Mₑ and fₑ are computed with stored material parameters and necessary assembly setups provided by the input PhaseSetup.

YiyuanStudentProject.compute_time_step! — Function

compute_time_step!(setup::RVESetup{dim}, load::LoadCase{dim}, Δt)

Return the object RVESetup with updated next time step solution vector aⁿ⁺¹.

Using Crank-Nicolson Method to compute the solution aⁿ⁺¹ in an implicit time stepping scheme for solving the linear system in the next time step.

New boundary condition values LoadCase are applied before for the next time step solutions.

YiyuanStudentProject.solve_time_series — Function

solve_time_series(rve::RVE{dim}, load::LoadCase{dim};  Δt=0.1, t_total=1) where {dim}

Return results throughout the time stepping process for visualizing the fine scale problem results.

The RVE problem is prepared, assembled and solved over a time series using the given RVE, LoadCase, t_total, Δt . By default t_total is set to 1 and Δt to 0.1.

Results are stored in a NamedTuple with fields total time t and solution vector a.

YiyuanStudentProject.plot_grid — Function

plot_grid(grid::Grid{3})

Ruturn a Makie.Figure to visualize a given 3D finite element grid using the Makie plotting library.

YiyuanStudentProject.animate_result — Function

animate_result(res::NamedTuple, setup::RVESetup{dim}, file_name::String="Myresult.mp4", kwargs...)

Return an mp4 file with name file_name, which by default is "Myresult.mp4".

This shows the evolution of the solution res for a 3D RVE simulation based on input RVESetup.

Further arguments like scale for visibel deformation can be used.