Types

YiyuanStudentProject.Material — Type

Material{dim,T}

A Material material type object used for defining the material character with the following parameters:

E: fourth order stiffness tensor E,

αᶜʰ: isotropic ion intercalation tensor,

k: concentration-chemical potantial coefficient,

cʳᵉᶠ: reference concentration,

M: mobility tensor,

μʳᵉᶠ reference chemical potantial.

The type can be created by calling the funtion Material{dim}(; G::T, K::T, η::T, cʳᵉᶠ::T, μʳᵉᶠ::T, θʳᵉᶠ::T, cᵐ::T, α::T, R::T=8.31446261815324) where {dim, T<:Real}

YiyuanStudentProject.RVE — Type

RVE{dim}

An object called RVE that contains

grid: a Grid for the prescribed RVE,

P: Material type for partical constraints,

M: Material type for matrix.

The type can be created by calling the funtion RVE(; grid::Grid{dim}, materials::NamedTuple{(:P,:M),Tuple{Material{dim}, Material{dim}}}) where {dim}

YiyuanStudentProject.LoadCase — Type

LoadCase{dim}

A LoadCase type object that defined the external forces

ε̄: averaging external strain Tensor,

μ̄: averaging chemical potantial on the boundary,

ζ̄ gradient of the averaging chemical potantial on the boundary.

The struct can be created by calling the funtion `LoadCase(dim::Int; kwargs...)

YiyuanStudentProject.PhaseSetup — Type

PhaseSetup{dim}

A PhaseSetup type object that contains the relevant imformations for the element assembly:

dh: dofHandler based on the given grid,

cells: ordered cell sets,

cv: cellvalues in named tuple for each unknown field representivly,

nbf: number of base function in named tuple for the unknown fields representivly,

material: material struct,

Kₑ: element stiffness matrix,

Mₑ: element mass matrix,

submatrices: submatrices for locating the unknown fields for element matrix assembly.

YiyuanStudentProject.RVESetup — Type

RVESetup{dim}

A PhaseSetup object that contains the relevant imformations for the problem solving/time stepping:

grid: a Grid for the prescribed RVE,

dh: dofHandler based on the given grid,

phasesetups: struct PhaseSetup in named tuple for each unknown field representivly,

K: stiffness matrix

M: mass matrix

J: jacobian matrix for time stepping

g: residual vector for time stepping

aⁿ: unknowns in current step

aⁿ⁺¹: unknowns in next step