Upscaling

YiyuanStudentProject.compute_effective_response! — Function

compute_effective_response!(gpdata::GaussPointData{dim}, rvesetup::RVESetup{dim}, load::LoadCase{dim}, Δt) where {dim}

Computes the varialtionally consistent macro scale fields from an RVE at a macro scale Gauss Point.

An RVE probelm is solved with given RVESetup and LoadCase for the new time step with time width Δt. State variables aᵣᵥₑⁿ, c̄ⁿ, c̄₂ⁿ from GaussPointData are updated.

YiyuanStudentProject.average_quantities — Function

average_quantities(a::Vector{Float64}, setup::RVESetup{dim}) where {dim}

Computes the average quantities σ̄, c̄, c̄₂, j̄ for a given solution vector a over the phase setups of an RVE.

Then the macro scale consistent macro scale fields are computed as following:

σ̄ = ∫ (E ⊡ (ε - αᶜʰ * (c - cʳᵉᶠ))) * dΩ / Vᵣᵥₑ

c̄ = ∫ c * dΩ / Vᵣᵥₑ

c̄₂ = ∫ (c * x) * dΩ / Vᵣᵥₑ

j̄ = - ∫ (M ⋅ ∇μ) * dΩ / Vᵣᵥₑ