Macro scale

YiyuanStudentProject.prepare_macro_setup — Function

prepare_macro_setup(grid::Grid{dim}, rvesetup::RVESetup{dim}, Δt) where {dim}

Return a SolveSetup object for the element-wise assembly and solving the time dependent macro scale problem.

Arguments:

grid: prescribed macro scale grid with no distinguished phases,

rvesetup: RVESetup,

Δt: The time step size.

Implementation Details:

The interpolation ip is defined by passing the corresponding refshape using the function _get_ref_shape(Val(dim)).

After the 'DofHandler' is created based on the grid, unknown fields can be added to it using add! By calling close! to finalize the construction.

Cellvallues are created for each discrete fields respectively passing quadrature rule with 3 integration points.

Number of base function for each unknown fields is defined using getnbasefunctions.

Calling function add_bc! to initialize the boundary conditions and associate this to the $Dofhandler$.

Local stiffness and solution vector are initialized. subarrays for locating the corresponding field interaction in element-wise by calling the macro @view.

Global stiffness matrix, global right-hand side and global solution vector are initialized. For the stiffness matrix using allocate_matrix for a sparse matrix pattern.

Create the necessary data storage $GaussPointData$ for each gauss point in macro scale with the initialized solution vector from RVESetup, 0.0 for c concentration and c₂ the gradient of concentration.

A $AssemblySetup$ for elementweise assembly and $SolveSetup$ for the final solving of the time dependent problem are prepared.

YiyuanStudentProject.add_macro_bc! — Function

addmacrobc!(ch::ConstraintHandler, grid::Grid{3}, Δt)

Add a time dependent dirichlet boundary condition for the macro scale unknown fields respectively on the possible facet ∂Ω1 and ∂Ω2.

Apply time dependent boundary condition on both unknown fields u and μ.

YiyuanStudentProject.assemble_macro_K! — Function

    assemble_macro_K!(setup::SolveSetup{dim}, Δt) where {dim}

Return the whole $SolveSetup$ with assembled stiffness matrix $K$.

Arguments:

setup: $SolveSetup$: A preconfigured object that defined for macro scale problem assembly. Fields from setup are used in this function:

assemblysetup: $AssemblySetup$: A preconfigured object type that defined for elementweise assembly,

K: Initialized global stiffness matrix to be assembled,

aⁿ: Initialized global solution vector for calling the element solution vector

Implementation Details:

Create a CSCAssembler for K.

Do the whole assembly: for all cells update the cell values for all the unknown fields and fill zeros to the do element assembly, aliening the element solution vector with the global solution vector. return assembled Kₑ.

YiyuanStudentProject.assemble_macro_element! — Function

    assemble_macro_element!(setup::AssemblySetup{dim}, Δt, gpdata::Vector{GaussPointData{dim}}) where {dim}

Return the assembled stiffness matrix.

Arguments:

setup: a preconfigured object type that defined for elementweise assembly.

Implementation Details:

The local unknowns μₑ and uₑ are passed for creating the step dependent macro scale variables for each quadrature point. Object $LoadCase$ is generated using the macro scale variables μ̄, ζ̄and ε̄. For each quadrature point the element volume is generated. The variationally consistent macro-scale (homogenized) fields σ̄, ċ, ċ₂, j̄ are computed passing the upscaling function $compute_effective_response!$. Furthermore for each base function in corresponding field, evaluate the shape function for the test function.

Then the coupling subarrays are computed as:

Kₑuu = ∫(δNϵi ⊡ σ̄ ) * dΩ

Kₑμμ = ∫(δNμi * ċ - δN∇μi ⋅ (ċ₂ - j̄) ) * dΩ

where:

δNxi: Shape function for x field test function

δN∇xi: Gradient of shape function for test function of field x

dΩ: Determinant of the Jacobian times the quadrature weight

YiyuanStudentProject.solve_macro_problem — Function

solve_macro_problem(grid::Grid{dim}, rvesetup::RVESetup{dim}; Δt=0.1, t_total=1) where {dim}

return solution storage and the macro scale setup for result visualizing. Compute the solution storage in a NamedTuple with fields t total time and a solution vector series contain u displacement and μ chemical potantial for the whole time series with a certain time step width.

Arguments

grid: Grid{dim}: The macro scale grid,

rvesetup: RVESetup{dim}: The object containing setups for solving rve problem,

Δt: The time step size,

t-total: the total computational time.

Implementation Details

Initialize the solution storage. Do time stepping with updating the boundary condition.

The updated boundary condition is then applied to the global stiffness matrix and force vector as well as the solution vector.

Using solver from Pardiso.jl for eventually ill conditioned system matrices. Update the boundary condition on the result vector.

Copy the new solution vector in the storage with the corresponding time step.

YiyuanStudentProject.animate_macro_result — Function

animate_macro_result(res::NamedTuple, setup::RVESetup{dim}, file_name::String="Myresult.mp4", n::Number)

Return an animation showing the evolution of the solution for a macro scale simulation.

Arguments:

res: A NamedTuple containing the simulation results

setup: The setup object for the RVE simulation, containing: grid and dh fields

file_name: The path and name of the output animation file (default: "Myresult.mp4")

Implementation Details:

Initialize a Makie figure with a size of (1200, 800).

Generate observables for t and a using _prepare_plots!.

Record frames by iterating over the indices of t and updates the observables accordingly.

Save the animation as an MP4 file.

YiyuanStudentProject._prepare_plots! — Function

prepareplots!(pos, res::NamedTuple, setup::RVESetup{dim}; scale::Real=1.0, titlestart::String="RVE response") where {dim}

Sets up the necessary plots and observables `tᵒᵇˢ` and `aᵒᵇˢ` for visualizing macroscale simulation results: deformation, chemical potential, concentration.

Arguments:

pos: A layout position,

res: A NamedTuple containing:

t: Time at each time step,

a: Result vector corresponding to the time.

setup: An object SolveSetupwith the grid and degrees of freedom configuration.

Implementation Details:

Generate the mesh representation both undeformed and deformed grid.

Determine the boundary values of chemical potential μ and concentration c across the grid nodes for all time steps.

Add sliced opening to show the internal strucature with particles and matrix with distinguished colors.

Create observables for time tᵒᵇˢ and data aᵒᵇˢ.

Prepare subplots for: Undeformed grid with a static wireframe. Deformed grid based on the scaled displacement field u. Chemical potential and concentration using a color-mapped mesh and color bar.

_prepare_plots!(pos, res::NamedTuple, setup::SolveSetup{dim}; 
                scale::Real=1.0, 
                titlestart::String="macroscale solution") where {dim}

Sets up the necessary plots and observables `tᵒᵇˢ` and `aᵒᵇˢ` for visualizing macroscale simulation results: deformation, chemical potential.

Arguments:

pos: A layout position,

res: A NamedTuple containing:

t: Time at each time step,

a: Result vector corresponding to the time.

setup: An object SolveSetupwith the grid and degrees of freedom configuration.

Implementation Details:

Generate the mesh representation both undeformed and deformed grid.

Determine the boundary values of chemical potential μ across the grid nodes for all time steps.

Create observables for time tᵒᵇˢ and data aᵒᵇˢ.

Prepare subplots for: Undeformed grid with a static wireframe. Deformed grid based on the scaled displacement field u. Chemical potential using a color-mapped mesh and color bar.

YiyuanStudentProject.animate_combined_result — Function

animate_macro_result(res::NamedTuple, setup::SolveSetup{dim}; file_name="Myresult.mp4", kwargs...) where {dim}

Generates an MP4 animation of a time dependent macro-scale results using Makie.jl.

Arguments:

res: A NamedTuple with the following fields:

t: Time at each time step,

a: Result vector corresponding to the time.

setup: An SolveSetup that contains the problem configuration.

Implementation Details:

Initialize a Makie figure with a size of (1200, 800).

Generate observables for t and a as well as the NamedTuple observables rve using _prepare_plots!.

Record frames by iterating over the indices of t and updates the observables accordingly.

Save the animation as an MP4 file.