Types

YiyuanStudentProject.Material — Type

Material{dim,T}

A $Material$ characterize the constitutive behavior with:

E: fourth order stiffness tensor E,

αᶜʰ: second order ion intercalation tensor,

k: concentration-chemical potential coefficient,

cʳᵉᶠ: reference concentration,

M: second order mobility tensor,

μʳᵉᶠ reference chemical potantial.

A $Material$ with isotropic properties can be created by calling the funtion Material{dim}(; G::T, K::T, η::T, cʳᵉᶠ::T, μʳᵉᶠ::T, θʳᵉᶠ::T, cᵐ::T, α::T, R::T=8.31446261815324) where {dim, T<:Real}

YiyuanStudentProject.RVE — Type

RVE{dim}

A $RVE$ is defined by a geometry and material characteristic of constituents that contains

grid: a Grid for the prescribed RVE,

P: Material type for partical constraints,

M: Material type for matrix.

An object can be created by calling the funtion RVE(; grid::Grid{dim}, materials::NamedTuple{(:P,:M),Tuple{Material{dim}, Material{dim}}}) where {dim}

YiyuanStudentProject.LoadCase — Type

LoadCase{dim}

A LoadCase defines the macro scale quantities imposed on RVE as loading

ε̄: average external strain tensor,

μ̄: average chemical potantial on the boundary,

ζ̄ gradient of the averaging chemical potantial on the boundary.

To create a $LoadCase$ by defining only the non-zero tensor components the constructor LoadCase(dim::Int; kwargs...) can be used. In case of a 2D problem, arguments like ε̄₁₁, ε̄₁₂, ε̄₂₂, μ̄, ζ̄₁, ζ̄₂ are prescribed. For 3D cases, the corresponding arguments are ε̄₁₁, ε̄₁₂, ε̄₁₃ ε̄₂₃, ε̄₂₂,ε̄₃₃,μ̄ ,ζ̄₁,ζ̄₂,ζ̄₃`.

YiyuanStudentProject.PhaseSetup — Type

PhaseSetup{dim}

A PhaseSetup contains the relevant imformation for the element assembly:

dh: $DofHandler$ based on the given grid,

cells: cell set defining the phase domain,

cv: $CellValues$ in $NamedTuple$ for each unknown field,

nbf: number of base function in $NamedTuple$ for the unknown fields,

material: $Material$ of the phase,

Kₑ: element stiffness matrix,

Mₑ: element mass matrix,

fₑ: element right hand side vector,

subarrays: subarrays blocks associated with the unknown fields for element matrix assembly.

YiyuanStudentProject.RVESetup — Type

RVESetup{dim}

A RVESetup contains the relevant imformation for the solving the transient problem:

grid: a $Grid$ defining the RVE geometry,

dh: $DofHandler$ based on the given grid,

phasesetups: $PhaseSetup$ in $NamedTuple$ for each unknown field,

K: stiffness matrix,

M: mass matrix,

f: right hand side vector,

J: system matrix for time stepping,

g: system vector for time stepping,

aⁿ: unknowns at current time,

aⁿ⁺¹: unknowns after performing a time,

Vʳᵛᵉ: Volume of the RVE.

YiyuanStudentProject.GaussPointData — Type

GaussPointData{dim}

A mutable GaussPointData that stores relevent data from the RVE problem for the one corresponding gauss point in macro scale problem:

aᵣᵥₑⁿ: the result vector from every rve problem,

c̄ⁿ: the average concentration at current time step,

c̄₂ⁿ: the average moment of concentration at current time step.

YiyuanStudentProject.AssemblySetup — Type

AssemblySetup{dim}

An AssemblySetup contains the relevant imformation for the element assembly in macro scale problem:

dh: $DofHandler$ based on the given grid,

cv: $CellValues$ in $NamedTuple$ for each unknown field,

nbf: number of base function in $NamedTuple$ for the unknown fields,

Kₑ: element stiffness matrix,

aₑ: element unknowns vector,

subarrays: subarrays blocks associated with the unknown fields for element matrix assembly,

gpdata: $GaussPointData$ at corresponding cell and gauss point,

rvesetup: $RVESetup$ for evaluating fine scale rve problem at each gauss point.

YiyuanStudentProject.SolveSetup — Type

SolveSetup{dim}

A SolveSetup contains the relevant imformation for the solving the transient macro scale problem:

grid: a $Grid$ defining the RVE geometry,

dh: $DofHandler$ based on the given grid,

ch: $ConstraintHandler$ for updating the boundary condition at each time step,

assemblysetup: $AssemblySetup$ for element assembly,

K: stiffness matrix,

f: right hand side vector,

aⁿ: unknowns at current time.

YiyuanStudentProject.MacroBCParams — Type

MacroBCParams

A MacroBCParams contains the possible boundary condition for the macro scale problem:

u_1: the total boundary condition for unknown field u on facet 1,

u_2: the total boundary condition for unknown field u on facet 2,

μ_1: the total boundary condition for unknown field μ on facet 1,

μ_2: the total boundary condition for unknown field μ on facet 2,

facetset_1: possible facet 1 for applying boundary condition,

facetset_2: possible facet 1 for applying boundary condition.